requirements.txt
Django>=5.0.0,<6.0.0
django-crispy-forms>=2.0
crispy-bootstrap5>=2023.1
Pillow>=10.0.0
python-dotenv>=1.0.0
whitenoise>=6.0.0
django-debug-toolbar>=4.0.0
django-extensions>=3.2.0
python-dateutil>=2.8.2
reportlab>=4.0.0
xhtml2pdf>=0.2.9
qrcode>=7.4.2
segno>=1.5.3
